refactor: dismember files registration-page and registration-form
