+++ src/feature/registration/fields/address-fields.tsx
@@ -4,7 +4,7 @@ import type { FieldErrors, UseFormRegister } from 'react-hook-form';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 
-import type { RegistrationFormData } from './types';
+import type { RegistrationFormData } from '../types';
 
 type AddressFieldsProps = {
   register: UseFormRegister<RegistrationFormData>;
