+++ src/feature/registration/fields/date-field.tsx
@@ -1,0 +1,37 @@
+import { AlertCircle } from 'lucide-react';
+import type { FieldErrors, UseFormRegister } from 'react-hook-form';
+
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+
+import type { RegistrationFormData } from '../types';
+
+type AddressFieldsProps = {
+  register: UseFormRegister<RegistrationFormData>;
+  errors: FieldErrors<RegistrationFormData>;
+};
+
+export function DateField({ register, errors }: AddressFieldsProps) {
+  return (
+    <>
+      <div className="grid gap-2">
+        <Label htmlFor="dateOfBirth">Date of Birth</Label>
+        <div className="relative">
+          <Input
+            id="dateOfBirth"
+            type="date"
+            {...register('dateOfBirth')}
+            max={new Date().toISOString().split('T')[0]}
+            aria-invalid={!!errors.dateOfBirth}
+          />
+          {errors.dateOfBirth && (
+            <AlertCircle className="absolute top-1/2 right-3 h-4 w-4 -translate-y-1/2 text-red-500" />
+          )}
+        </div>
+        {errors.dateOfBirth && (
+          <p className="error-message text-sm text-red-500">{errors.dateOfBirth.message}</p>
+        )}
+      </div>
+    </>
+  );
+}
